+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Snapshots now cannot be created while SkipDisk is active
+- Autoplacer: Added very small default weight for the MinResourceCount-strategy for better tiebreaker-distribution
 
 ### Fixed